requirements.txt
aiohttp==3.8.4
aiohttp-apispec==2.2.3
aiohttp-session==2.12.0
aiosignal==1.3.1
apispec==3.3.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
multidict==6.0.4
packaging==23.0
pycparser==2.21
python-dotenv==1.0.0
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
webargs==5.5.3
yarl==1.8.2
